backend/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
flask_login==0.6.2
flask-SQLAlchemy==3.1.1
werkzeug==2.3.7
Flask-Cors==4.0.0
mysqlclient==2.2.0
gunicorn==21.2.0
